# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WorkClub API
  version: 1.1.0
  description: |
    Multi-tenant API for agency client work, projects, tasks, time, proposals,
    invoices, teams, notifications and the restricted client portal.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://localhost:8888/api
    description: Local development
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200': { $ref: '#/components/responses/AuthSession' }
        '401': { $ref: '#/components/responses/Error' }
  /auth/register-workspace:
    post:
      tags: [Authentication]
      security: []
      operationId: registerWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201': { $ref: '#/components/responses/AuthSession' }
        '409': { $ref: '#/components/responses/Error' }
  /auth/accept-invite:
    post:
      tags: [Authentication]
      security: []
      operationId: acceptInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, name, password]
              properties:
                token: { type: string }
                name: { type: string }
                password: { type: string, format: password }
      responses:
        '201': { $ref: '#/components/responses/AuthSession' }
        '410': { $ref: '#/components/responses/Error' }
  /auth/refresh:
    post:
      tags: [Authentication]
      security: []
      operationId: refreshSession
      description: Rotates the HttpOnly refresh cookie and returns a short-lived access token.
      responses:
        '200': { $ref: '#/components/responses/AuthSession' }
        '401': { $ref: '#/components/responses/Error' }
  /auth/request-password-reset:
    post:
      tags: [Authentication]
      security: []
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202': { $ref: '#/components/responses/Success' }
  /auth/reset-password:
    post:
      tags: [Authentication]
      security: []
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, format: password }
      responses:
        '200': { $ref: '#/components/responses/AuthSession' }
        '410': { $ref: '#/components/responses/Error' }
  /auth/verify-email:
    post:
      tags: [Authentication]
      security: []
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200': { $ref: '#/components/responses/AuthSession' }
        '410': { $ref: '#/components/responses/Error' }
...
  /auth/me:
    get:
      tags: [Authentication]
      operationId: currentSession
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
...
  /auth/sessions/{sessionId}:
    delete:
      tags: [Authentication]
      operationId: revokeSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    AuthSession:
      description: Authentication succeeded
      headers:
        Set-Cookie:
          description: Rotating HttpOnly refresh cookie
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    RegisterInput:
      type: object
      additionalProperties: false
      required: [workspaceName, name, email, password]
      properties:
        workspaceName: { type: string, minLength: 2, maxLength: 80 }
        name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 10 }
    SuccessEnvelope:
      type: object
      required: [success, result]
      properties:
        success: { const: true }
        result: {}
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { const: false }
        message: { type: string }
        details: {}
...
security:
  - bearerAuth: []